# docs/swagger.yaml
definitions:
...
  usercenter.CommonResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  usercenter.DelRequest:
    properties:
      id:
        description: '@inject_tag: validate:"required"'
        type: string
      tenantID:
        type: string
    required:
    - id
    type: object
...
  usercenter.GetProjectDetailResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/usercenter.ProjectInfo'
      message:
        type: string
    type: object
...
  usercenter.ProjectFormComponent:
    properties:
      id:
        type: string
      name:
        type: string
      projectID:
        type: string
    type: object
  usercenter.ProjectInfo:
    properties:
      cellCount:
        type: integer
      description:
        type: string
      expired:
        type: string
      formComponents:
        items:
          $ref: '#/definitions/usercenter.ProjectFormComponent'
        type: array
      formCount:
        type: integer
      id:
        type: string
      isMust:
        description: 系统必须要有的数据
        type: boolean
      name:
        type: string
      pageCount:
        type: integer
      tenantID:
        type: string
    type: object
...
  usercenter.QueryProjectResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/usercenter.ProjectInfo'
        type: array
      message:
        type: string
      pages:
        type: integer
      records:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/core/auth/project/add:
    post:
      consumes:
      - application/json
      description: 新增
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Add Project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.ProjectInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 新增
      tags:
      - 项目管理
  /api/core/auth/project/all:
    get:
      consumes:
      - application/json
      description: 查询所有
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryProjectResponse'
      summary: 查询所有
      tags:
      - 项目管理管理
  /api/core/auth/project/delete:
    delete:
      consumes:
      - application/json
      description: 删除
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Delete Project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.DelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 删除
      tags:
      - 项目管理
  /api/core/auth/project/detail:
    get:
      consumes:
      - application/json
      description: 查询明细
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.GetProjectDetailResponse'
      summary: 查询明细
      tags:
      - 项目管理
  /api/core/auth/project/export:
    get:
      consumes:
      - application/json
      description: 导出
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 租户ID
        in: query
        name: tenantID
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      - description: IDs
        in: query
        items:
          type: string
        name: ids
        type: array
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导出
      tags:
      - 项目管理管理
  /api/core/auth/project/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入
      parameters:
      - description: Bearer+空格+Token
        in: header
        name: authorization
        required: true
        type: string
      - description: 要上传的文件
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 导入
      tags:
      - 项目管理管理
  /api/core/auth/project/query:
    get:
      consumes:
      - application/json
      description: 分页查询
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: 从1开始
        in: query
        name: pageIndex
        type: integer
      - description: 默认每页10条
        in: query
        name: pageSize
        type: integer
      - description: 排序字段
        in: query
        name: orderField
        type: string
      - description: 是否倒序排序
        in: query
        name: desc
        type: boolean
      - description: 租户ID
        in: query
        name: tenantID
        type: string
      - description: 名称
        in: query
        name: name
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.QueryProjectResponse'
      summary: 分页查询
      tags:
      - 项目管理
  /api/core/auth/project/update:
    put:
      consumes:
      - application/json
      description: 更新
      parameters:
      - description: jwt token
        in: header
        name: authorization
        required: true
        type: string
      - description: Update Project
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usercenter.ProjectInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usercenter.CommonResponse'
      summary: 更新
      tags:
      - 项目管理
...
swagger: "2.0"